requirements.txt
Flask==2.3.3
flask-mysqldb==1.0.1
flask-socketio==5.3.4
flask-session==0.5.0
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
python-socketio==5.9.0
python-engineio==4.7.1
mysqlclient==2.2.0
redis==5.0.1
qrcode==7.4.2
Pillow==10.1.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0